Split commands tests
